[options] Move ClusterOption into sos.options
Moves `ClusterOption` into sos.options, instead of living in
sos.collector.clusters.
